Extract display_assignments to seatrades/visualization.py (#55)
