+++ seatrades/results.py
@@ -1,12 +1,9 @@
-"""
-This file contains tools to display the results of seatrades assignment.
-"""
+"""Result data structures and wrangling functions for seatrade assignments."""
 
 from dataclasses import dataclass
 from enum import Enum
 from typing import Optional
 
-import altair as alt
 import pandas as pd
 
 
@@ -47,9 +44,6 @@ class AssignmentSolution:
     seatrades_full: list[str]
     cabin_camper_prefs: pd.DataFrame
     camper_prefs: pd.Series
-
-
-alt.data_transformers.disable_max_rows()
 
 
 def wrangle_assignments_to_longform(solution: AssignmentSolution) -> pd.DataFrame:
@@ -149,44 +143,3 @@ def prepare_seatrade_leaders(longform_df: pd.DataFrame) -> pd.DataFrame:
     assigned = longform_df[longform_df["assignment"] == 1.0]
     sorted_assigned = assigned.sort_values(by=["block", "seatrade", "cabin", "camper"], kind="stable")
     return sorted_assigned[["block", "seatrade", "camper", "cabin"]].reset_index(drop=True)
-
-
-def display_assignments(solution: AssignmentSolution) -> alt.Chart:
-    """Display the assignments of the seatrades visually for inference."""
-    if solution.status.state == SolverState.ERROR:
-        raise ValueError(f"No solution found. {solution.status.message}")
-    elif solution.status.state == SolverState.INFEASIBLE:
-        raise ValueError(
-            f"Solver status ({solution.status.state.value}) indicates "
-            "the problem was not successfully solved. Refusing to render untrustworthy results."
-        )
-
-    df = wrangle_assignments_to_longform(solution)
-
-    assignment_base = alt.Chart(df).encode(
-        x=alt.X("seatrade", sort=solution.seatrades_full, title=None),
-        y=alt.Y("camper", sort=solution.campers, title=None),
-    )
-    assignment_rectangles = assignment_base.mark_rect(stroke="black", strokeWidth=0.1).encode(
-        color=alt.Color(
-            "preference:O",
-        )
-    )
-    assignment_text = (
-        assignment_base.mark_text(color="white").encode(text="preference:O").transform_filter(alt.datum.preference > 0)
-    )
-    assignment_chart = (
-        (assignment_rectangles + assignment_text)
-        .facet(row="cabin", column="block", spacing={"row": 2})
-        .resolve_scale(y="independent")
-        .properties(
-            title={
-                "text": "Seatrades.",
-                "subtitle": "Assignments by Preference.",
-                "fontSize": 20,
-                "anchor": "start",
-            }
-        )
-    )
-
-    return assignment_chart
