+++ tests/test_seatrades/test_results.py
@@ -1,14 +1,10 @@
 """Tests for seatrades/results.py."""
 
-import dataclasses
-
 import pandas as pd
-import pytest
 
 from seatrades.results import (
     SolverState,
     SolverStatus,
-    display_assignments,
     wrangle_assignments_to_longform,
 )
 
@@ -140,30 +136,3 @@ class TestWrangleAssignmentsToLongform:
         result = wrangle_assignments_to_longform(sample_assignment_solution)
         assigned = result[result["assignment"] == 1.0]
         assert set(assigned["seatrade"].unique()) == {"Archery", "Sailing", "Climbing"}
-
-
-class TestDisplayAssignmentsFailureGuard:
-    def test_raises_on_infeasible_status(self, sample_assignment_solution):
-        """display_assignments must raise ValueError when optimization was infeasible."""
-
-        infeasible = dataclasses.replace(
-            sample_assignment_solution,
-            status=SolverStatus(state=SolverState.INFEASIBLE),
-        )
-        with pytest.raises(ValueError, match="not successfully solved"):
-            display_assignments(infeasible)
-
-    def test_raises_on_error_status(self, sample_assignment_solution):
-        """display_assignments must raise ValueError when status is ERROR (unsolved)."""
-
-        error_solution = dataclasses.replace(
-            sample_assignment_solution,
-            status=SolverStatus(state=SolverState.ERROR, message="Not solved"),
-        )
-        with pytest.raises(ValueError, match="No solution found"):
-            display_assignments(error_solution)
-
-    def test_returns_chart_on_optimal(self, sample_assignment_solution):
-        """display_assignments returns a chart when optimization succeeded."""
-        result = display_assignments(sample_assignment_solution)
-        assert result is not None
