+++ tests/test_seatrades/test_visualization.py
@@ -1,0 +1,35 @@
+"""Tests for seatrades/visualization.py."""
+
+import dataclasses
+
+import pytest
+
+from seatrades.results import SolverState, SolverStatus
+from seatrades.visualization import display_assignments
+
+
+class TestDisplayAssignmentsFailureGuard:
+    def test_raises_on_infeasible_status(self, sample_assignment_solution):
+        """display_assignments must raise ValueError when optimization was infeasible."""
+
+        infeasible = dataclasses.replace(
+            sample_assignment_solution,
+            status=SolverStatus(state=SolverState.INFEASIBLE),
+        )
+        with pytest.raises(ValueError, match="not successfully solved"):
+            display_assignments(infeasible)
+
+    def test_raises_on_error_status(self, sample_assignment_solution):
+        """display_assignments must raise ValueError when status is ERROR (unsolved)."""
+
+        error_solution = dataclasses.replace(
+            sample_assignment_solution,
+            status=SolverStatus(state=SolverState.ERROR, message="Not solved"),
+        )
+        with pytest.raises(ValueError, match="No solution found"):
+            display_assignments(error_solution)
+
+    def test_returns_chart_on_optimal(self, sample_assignment_solution):
+        """display_assignments returns a chart when optimization succeeded."""
+        result = display_assignments(sample_assignment_solution)
+        assert result is not None
